added reservation handlebar partial
