+++ routes/api-routes.js
@@ -2,41 +2,6 @@
 var db = require("../models");
 var passport = require("../config/passport");
 
-async function viewAllReservations(db) {
-  let results;
-  try {
-    results = await db.reservations.findAll({
-      include: [
-        {
-          model: db.users
-        },
-        {
-          model: db.users,
-          as: "instructor"
-        }
-      ]
-    });
-  } catch (err) {
-    console.log("Something went sideways", err);
-  }
-
-  let values = [];
-  results.forEach(res => {
-    values.push({
-      id: res.id,
-      start: res.start_Time,
-      end: res.end_Time,
-      instructor: res.instructor.first_name + " " + res.instructor.last_name,
-      customer: res.user.first_name + " " + res.user.last_name,
-      email: res.user.email,
-      status: res.status,
-      reservation_number: res.session_ID
-    });
-  });
-
-  return values;
-}
-
 module.exports = function(app) {
   // Using the passport.authenticate middleware with our local strategy.
   // If the user has valid login credentials, send them to the members page.
