+++ views/members.handlebars
@@ -26,7 +26,9 @@
       </div>
       <div class="row">
         <div class="col-md-6 col-md-offset-3">
+          {{#if event}}
           {{> reserved}}
+          {{/if}}
         </div>
       </div>
   </div>
