+++ views/partials/reserved.handlebars
@@ -3,7 +3,7 @@
 <ul>
     <li class="cd-schedule__event">
         <a data-start="{{eventStart}}" data-end="{{eventEnd}}"  data-content="{{eventType}}" data-event="event-3" href="#0">
-            <em class="cd-schedule__name">{{event}}</em>
+            <em class="cd-schedule__name">{{dayoftheweek}} : {{event}}</em>
         </a>
     </li>
 </ul>
